--- imp_homework_on_swift/imp_homework_on_swift/StyledRoundedImage.swift
@@ -1,28 +1,0 @@
-//
-//  StyledRoundedImage.swift
-//  imp_homework_on_swift
-//
-//  Created by Татьяна  Травкина on 04.04.2021.
-//
-
-import UIKit
-
-class StyledRoundedImage: UIImageView {
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        self.setView()
-    }
-    
-    required init?(coder: NSCoder) {
-        super.init(coder: coder)
-        self.setView()
-    }
-    
-    private func setView() {
-        self.layer.borderWidth = 3.0
-        self.layer.borderColor = UIColor.white.cgColor
-        self.layer.cornerRadius = self.bounds.size.width / 2.0
-        self.clipsToBounds = true
-    }
-}
