+++ imp_homework_on_swift/imp_homework_on_swift/Views/Planet/PlanetTableViewCell.xib
@@ -1,0 +1,73 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="17701" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+    <device id="retina6_1" orientation="portrait" appearance="light"/>
+    <dependencies>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="17703"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <tableViewCell clipsSubviews="YES" contentMode="scaleToFill" preservesSuperviewLayoutMargins="YES" selectionStyle="default" indentationWidth="0.0" rowHeight="96" id="zZZ-45-wH5" customClass="PlanetTableViewCell" customModule="imp_homework_on_swift" customModuleProvider="target">
+            <rect key="frame" x="0.0" y="0.0" width="402" height="96"/>
+            <autoresizingMask key="autoresizingMask"/>
+            <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" preservesSuperviewLayoutMargins="YES" insetsLayoutMarginsFromSafeArea="NO" tableViewCell="zZZ-45-wH5" id="kS9-Af-3hF">
+                <rect key="frame" x="0.0" y="0.0" width="402" height="96"/>
+                <autoresizingMask key="autoresizingMask"/>
+                <subviews>
+                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="7QS-iC-Sva" userLabel="Location Label">
+                        <rect key="frame" x="20" y="26" width="191" height="21"/>
+                        <fontDescription key="fontDescription" type="system" pointSize="17"/>
+                        <nil key="textColor"/>
+                        <nil key="highlightedColor"/>
+                    </label>
+                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="right" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="Zaw-6N-fk3" userLabel="Population Label">
+                        <rect key="frame" x="238" y="-46" width="86" height="188"/>
+                        <fontDescription key="fontDescription" type="system" pointSize="17"/>
+                        <nil key="textColor"/>
+                        <nil key="highlightedColor"/>
+                    </label>
+                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="jNb-fu-MNN" userLabel="Type Location Label">
+                        <rect key="frame" x="20" y="55" width="191" height="14"/>
+                        <fontDescription key="fontDescription" type="system" pointSize="17"/>
+                        <nil key="textColor"/>
+                        <nil key="highlightedColor"/>
+                    </label>
+                    <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="D70-iT-fQh" userLabel="More Button">
+                        <rect key="frame" x="345" y="39" width="30" height="18"/>
+                        <state key="normal" backgroundImage="greaterthan" catalog="system"/>
+                        <connections>
+                            <action selector="moreButton:" destination="zZZ-45-wH5" eventType="touchUpInside" id="9YF-3E-Bqx"/>
+                        </connections>
+                    </button>
+                </subviews>
+                <constraints>
+                    <constraint firstAttribute="trailing" secondItem="7QS-iC-Sva" secondAttribute="trailing" constant="191" id="23T-4X-Yce"/>
+                    <constraint firstItem="Zaw-6N-fk3" firstAttribute="top" secondItem="kS9-Af-3hF" secondAttribute="topMargin" constant="-57" id="44g-ru-27A"/>
+                    <constraint firstItem="7QS-iC-Sva" firstAttribute="top" secondItem="kS9-Af-3hF" secondAttribute="top" constant="26" id="9OR-5R-Nqh"/>
+                    <constraint firstItem="D70-iT-fQh" firstAttribute="leading" secondItem="Zaw-6N-fk3" secondAttribute="trailing" constant="21" id="CkV-m1-9ob"/>
+                    <constraint firstAttribute="trailingMargin" secondItem="D70-iT-fQh" secondAttribute="trailing" constant="7" id="F2X-tq-jxP"/>
+                    <constraint firstItem="Zaw-6N-fk3" firstAttribute="centerY" secondItem="D70-iT-fQh" secondAttribute="centerY" id="FGW-Ms-XgT"/>
+                    <constraint firstItem="Zaw-6N-fk3" firstAttribute="leading" secondItem="kS9-Af-3hF" secondAttribute="leading" constant="238" id="UR2-y8-hB8"/>
+                    <constraint firstAttribute="trailing" secondItem="Zaw-6N-fk3" secondAttribute="trailing" constant="78" id="Vmk-MY-scF"/>
+                    <constraint firstItem="jNb-fu-MNN" firstAttribute="top" secondItem="kS9-Af-3hF" secondAttribute="top" constant="55" id="YJL-Gk-qFW"/>
+                    <constraint firstItem="Zaw-6N-fk3" firstAttribute="centerY" secondItem="kS9-Af-3hF" secondAttribute="centerY" id="dD6-FY-CUr"/>
+                    <constraint firstAttribute="trailing" secondItem="jNb-fu-MNN" secondAttribute="trailing" constant="191" id="eRt-QD-tUW"/>
+                    <constraint firstItem="jNb-fu-MNN" firstAttribute="leading" secondItem="kS9-Af-3hF" secondAttribute="leading" constant="20" symbolic="YES" id="esK-Hi-uYb"/>
+                    <constraint firstAttribute="bottom" secondItem="7QS-iC-Sva" secondAttribute="bottom" constant="49" id="hsU-PE-JoL"/>
+                    <constraint firstAttribute="bottom" secondItem="jNb-fu-MNN" secondAttribute="bottom" constant="27" id="mNK-mL-DKZ"/>
+                    <constraint firstItem="7QS-iC-Sva" firstAttribute="leading" secondItem="kS9-Af-3hF" secondAttribute="leading" constant="20" symbolic="YES" id="wai-uC-J08"/>
+                </constraints>
+            </tableViewCellContentView>
+            <connections>
+                <outlet property="LocationLabel" destination="7QS-iC-Sva" id="d7E-Ze-vES"/>
+                <outlet property="PopulationLabel" destination="Zaw-6N-fk3" id="6Ue-ck-Rup"/>
+                <outlet property="TypeLocationLabel" destination="jNb-fu-MNN" id="2SV-Qf-Q4C"/>
+            </connections>
+            <point key="canvasLocation" x="230.43478260869566" y="36.160714285714285"/>
+        </tableViewCell>
+    </objects>
+    <resources>
+        <image name="greaterthan" catalog="system" width="128" height="111"/>
+    </resources>
+</document>
