+++ imp_homework_on_swift/imp_homework_on_swift/Views/Profile/TopTableViewCell.xib
@@ -1,0 +1,80 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="17701" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+    <device id="retina6_1" orientation="portrait" appearance="light"/>
+    <dependencies>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="17703"/>
+        <capability name="System colors in document resources" minToolsVersion="11.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <tableViewCell clipsSubviews="YES" contentMode="scaleToFill" preservesSuperviewLayoutMargins="YES" selectionStyle="default" indentationWidth="10" rowHeight="400" id="Bod-Rj-eU8" customClass="TopTableViewCell" customModule="imp_homework_on_swift" customModuleProvider="target">
+            <rect key="frame" x="0.0" y="0.0" width="414" height="401"/>
+            <autoresizingMask key="autoresizingMask"/>
+            <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" preservesSuperviewLayoutMargins="YES" insetsLayoutMarginsFromSafeArea="NO" tableViewCell="Bod-Rj-eU8" id="A3T-nn-v32">
+                <rect key="frame" x="0.0" y="0.0" width="414" height="401"/>
+                <autoresizingMask key="autoresizingMask"/>
+                <subviews>
+                    <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="PKZ-g0-Rn7" userLabel="TopView">
+                        <rect key="frame" x="0.0" y="0.0" width="414" height="200"/>
+                        <subviews>
+                            <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleToFill" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="xHw-FT-Cmn" userLabel="Top Image View">
+                                <rect key="frame" x="0.0" y="0.0" width="414" height="200"/>
+                            </imageView>
+                        </subviews>
+                        <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+                        <constraints>
+                            <constraint firstItem="xHw-FT-Cmn" firstAttribute="top" secondItem="PKZ-g0-Rn7" secondAttribute="top" id="0lb-QH-oyp"/>
+                            <constraint firstAttribute="trailing" secondItem="xHw-FT-Cmn" secondAttribute="trailing" id="3Vh-Vn-jCU"/>
+                            <constraint firstAttribute="bottom" secondItem="xHw-FT-Cmn" secondAttribute="bottom" id="dmQ-9L-VOc"/>
+                            <constraint firstItem="xHw-FT-Cmn" firstAttribute="leading" secondItem="PKZ-g0-Rn7" secondAttribute="leading" id="pc7-vN-QlU"/>
+                        </constraints>
+                    </view>
+                    <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="9uf-Ib-BMm" userLabel="BottomView">
+                        <rect key="frame" x="0.0" y="201" width="414" height="200"/>
+                        <subviews>
+                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" fixedFrame="YES" text="Label" textAlignment="center" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="flf-7u-f7I">
+                                <rect key="frame" x="8" y="138" width="398" height="43"/>
+                                <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+                                <fontDescription key="fontDescription" type="system" pointSize="36"/>
+                                <color key="textColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+                                <nil key="highlightedColor"/>
+                            </label>
+                        </subviews>
+                        <color key="backgroundColor" red="0.5647850633" green="0.56485521790000004" blue="0.56476122139999996" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+                    </view>
+                    <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" ambiguous="YES" translatesAutoresizingMaskIntoConstraints="NO" id="gRo-ir-Efz" customClass="StyledRoundedImage" customModule="imp_homework_on_swift" customModuleProvider="target">
+                        <rect key="frame" x="97" y="91" width="219" height="219"/>
+                        <constraints>
+                            <constraint firstAttribute="width" secondItem="gRo-ir-Efz" secondAttribute="height" multiplier="1:1" id="Na9-8w-rBn"/>
+                        </constraints>
+                    </imageView>
+                </subviews>
+                <constraints>
+                    <constraint firstAttribute="bottom" secondItem="9uf-Ib-BMm" secondAttribute="bottom" id="3SX-EQ-WTn"/>
+                    <constraint firstAttribute="bottom" secondItem="PKZ-g0-Rn7" secondAttribute="bottom" constant="201" id="4Vb-6r-yfs"/>
+                    <constraint firstAttribute="trailing" secondItem="9uf-Ib-BMm" secondAttribute="trailing" id="FnE-sp-QiV"/>
+                    <constraint firstItem="9uf-Ib-BMm" firstAttribute="leading" secondItem="A3T-nn-v32" secondAttribute="leading" id="GDb-yb-xdG"/>
+                    <constraint firstAttribute="trailing" secondItem="PKZ-g0-Rn7" secondAttribute="trailing" id="ILZ-6A-eFh"/>
+                    <constraint firstItem="gRo-ir-Efz" firstAttribute="centerY" secondItem="A3T-nn-v32" secondAttribute="centerY" id="KdO-qQ-4Ei"/>
+                    <constraint firstItem="gRo-ir-Efz" firstAttribute="centerX" secondItem="A3T-nn-v32" secondAttribute="centerX" id="VQD-Gh-hJX"/>
+                    <constraint firstItem="PKZ-g0-Rn7" firstAttribute="top" secondItem="A3T-nn-v32" secondAttribute="top" id="n7d-91-cS1"/>
+                    <constraint firstItem="9uf-Ib-BMm" firstAttribute="top" secondItem="A3T-nn-v32" secondAttribute="top" constant="201" id="wzi-kb-kYv"/>
+                    <constraint firstItem="PKZ-g0-Rn7" firstAttribute="leading" secondItem="A3T-nn-v32" secondAttribute="leading" id="zfw-PE-2OR"/>
+                </constraints>
+            </tableViewCellContentView>
+            <connections>
+                <outlet property="TopImageView" destination="xHw-FT-Cmn" id="8h5-fD-ivh"/>
+                <outlet property="UserLoginLabel" destination="flf-7u-f7I" id="JKt-2j-5Nm"/>
+                <outlet property="avatarImage" destination="gRo-ir-Efz" id="22w-hj-5oU"/>
+            </connections>
+            <point key="canvasLocation" x="237.68115942028987" y="142.29910714285714"/>
+        </tableViewCell>
+    </objects>
+    <resources>
+        <systemColor name="systemBackgroundColor">
+            <color white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+        </systemColor>
+    </resources>
+</document>
